offer statede - bugs
 stil bug with product list

news delete dont work

+++ news/newsCreateView.php
@@ -1,6 +1,39 @@
 <?php
 require_once("../admin/adminHeader.php");
 require_once("NewsController.php");
+require_once("../dailySpecial/DailySpecialController.php");
+
+$dailyCon = new DailySpecialController();
+
+if (isset($_SESSION["discountProducts"])) {
+    $dailyCon->existingDiscountProduct($_SESSION["discountProducts"]);
+}
+
+if (!empty($_GET["action"])) {
+    if (isset($_GET['ID'])) {
+        $productID = $_GET['ID'];
+    }
+
+    //start the switch/case
+    switch ($_GET["action"]) {
+            //adding items to cart
+        case "remove":
+
+            $dailyCon->removeDiscountProduct($productID);
+
+            $adminId =  $_SESSION['admin_id'];
+            $admin = $_SESSION['admin'];
+
+            $_SESSION = $dailyCon->productArray;
+
+            $_SESSION['admin_id'] = $adminId;
+            $_SESSION['admin'] = $admin;
+            break;
+        case "empty":
+            unset($_SESSION["discountProducts"]);
+            break;
+    }
+}
 
 
 ?>
@@ -11,7 +44,7 @@ require_once("NewsController.php");
         <div class="row" style="margin-top:40px;">
 
             <h2>Create News Post</h2>
-            <form method="post">¨
+            <form method="post">
 
                 <div class="col s12">
                     <div class="col s6">
@@ -37,7 +70,7 @@ require_once("NewsController.php");
                             <a href="newsOverView.php" class="waves-effect waves-light btn red">Cancel</a>
                         </div>
                     </div>
-                    
+
                     <div class="col s6">
                         <a class="btn" href="../dailySpecial/addDiscoutToProductView.php">Add Product to Discount</a>
                         <table class="highlight">
@@ -45,27 +78,26 @@ require_once("NewsController.php");
                                 <tr>
                                     <th>ID</th>
                                     <th>Name</th>
-                                    <th>Color</th>
                                     <th>Price</th>
                                     <th>Quantity</th>
                                     <th>Delete</th>
                                 </tr>
                             </thead>
-
-                            <tbody>
-                                <?php
-                                foreach ($getProducts as $getProduct) {
-                                    echo "<tr>";
-                                    echo "<td>" . $getProduct['productID'] . "</td>";
-                                    echo "<td>" . $getProduct['name'] . "</td>";
-                                    echo "<td>" . $getProduct['color'] . "</td>";
-                                    echo "<td>" . $getProduct['price'] . "</td>";
-                                    echo "<td>" . $getProduct['quantity'] . "</td>";
-                                    echo '<td><a href="' . $getProduct['productID'] . '" class="waves-effect waves-light btn red" onclick="return confirm(\'Remove! are you sure?\')">remove</a></td>';
-                                    echo "</tr>";
-                                }
-                                ?>
-                            </tbody>
+                            <?php if (isset($_SESSION["discountProducts"])) { ?>
+                                <tbody>
+                                    <?php
+                                    foreach ($_SESSION["discountProducts"] as $getProduct) {
+                                        echo "<tr>";
+                                        echo "<td>" . $getProduct['productID'] . "</td>";
+                                        echo "<td>" . $getProduct['name'] . "</td>";
+                                        echo "<td>" . $getProduct['price'] . "</td>";
+                                        echo "<td>" . $getProduct['quantity'] . "</td>";
+                                        echo '<td><a href="newsCreateView.php?action=remove&ID=' . $getProduct['productID'] . '" class="btn red" onclick="return confirm(\'Remove! are you sure?\')">Remove</a></td>';
+                                        echo "</tr>";
+                                    }
+                                    ?>
+                                </tbody>
+                            <?php } ?>
                         </table>
                     </div>
                 </div>
@@ -80,7 +112,7 @@ require_once("NewsController.php");
 <?php
 if (isset($_POST['submit'])) { // Form has been submitted.
     $newsCon = new NewsController();
-    $newsCon->manangeCreate($_POST["title"], $_POST["desc"], $_POST["date"], $_POST["discount"]);
+    $newsCon->manangeCreate($_POST["title"], $_POST["desc"], $_POST["date"], $_POST["discount"], $_SESSION["discountProducts"]);
 }
 
 ?>
